fixbug

+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.dom/src/main/java/nts/uk/ctx/at/shared/dom/remainingnumber/paymana/PayoutManagementDataService.java
@@ -263,7 +263,7 @@ public class PayoutManagementDataService {
 			// Update remain days 振出管理データ
 			Optional<PayoutManagementData> payoutMan = payoutManagementDataRepository.findByID(item.getPayoutId());
 			if (payoutMan.isPresent()) {
-				payoutMan.get().setRemainNumber(Double.valueOf(item.getUsedDays().v().intValue()));
+				payoutMan.get().setRemainNumber(item.getUsedDays().v().doubleValue());
 				payoutMan.get().setStateAtr(DigestionAtr.UNUSED.value);
 				payoutManagementDataRepository.update(payoutMan.get());
 			}
